use long instead of integer

+++ dev/com.ibm.ws.sipcontainer/src/com/ibm/ws/sip/container/osgi/ServletInstanceHolderInterface.java
@@ -38,7 +38,7 @@ public interface ServletInstanceHolderInterface {
 	/**
 	 * Trigger listener
 	 */
-	public void triggerSipletInitServlet(int appQueueIndex);
+	public void triggerSipletInitServlet(long appQueueIndex);
 	
 	
 	/**
